Move `ServiceList` to `Icinga\Module\Icingadb\Widget\ItemList`
